GIX-2060: Remove mainTransactionFeeStore (#4308)
# Motivation

Use `bigint` for e8s and ulps.

In this PR, Remove deprecated store `mainTransactionFeeStore`.

The reason to stop using it is because it returned e8s in the `number`
type instead of `bigint`.

# Changes

* Remove `mainTransactionFeeStore` and use instead
`mainTransactionFeeE8sStore`.
* Convert `MIN_NEURON_STAKE` to `bigint`.
* Change parameter and return type in `minNeuronSplittable` to `bigint`.
* Change parameter type in `neuronCanBeSplit` to `bigint`.

# Tests

* Remove test for `mainTransactionFeeStore`.
* Convert parameter inputs from number to bigint.

# Todos

- [ ] Add entry to changelog (if necessary).

Not necessary.

+++ frontend/src/tests/lib/services/neurons.services.spec.ts
@@ -1883,7 +1883,7 @@ describe("neurons-services", () => {
           ...mockNeuron,
           fullNeuron: {
             ...mockFullNeuron,
-            cachedNeuronStake: BigInt(MIN_NEURON_STAKE / 2 - 10),
+            cachedNeuronStake: MIN_NEURON_STAKE / 2n - 10n,
           },
         },
         amount: 0.01,
